Toto/trunk/src/com/azquo/toto/service/LabelEditorService.java
package com.azquo.toto.service;

import com.azquo.toto.dao.LabelDAO;
import com.azquo.toto.entity.Label;
import org.springframework.beans.factory.annotation.Autowired;

/**
 * Created with IntelliJ IDEA.
 * User: cawley
 * Date: 17/10/13
 * Time: 14:18
 * I've been convinced that a service layer is probably a good idea, with the DAO/Database it can form the model.
 */
public class LabelEditorService {

    @Autowired
    private LabelDAO labelDAO;

    public Label findByName(String name){
        return labelDAO.findByName(name);
    }
}
